NIFI-3455 Large row count paging
This closes #1499.

+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/main/java/org/apache/nifi/processors/standard/db/impl/OracleDatabaseAdapter.java
@@ -29,7 +29,7 @@ public class OracleDatabaseAdapter implements DatabaseAdapter {
     }
 
     @Override
-    public String getSelectStatement(String tableName, String columnNames, String whereClause, String orderByClause, Integer limit, Integer offset) {
+    public String getSelectStatement(String tableName, String columnNames, String whereClause, String orderByClause, Long limit, Long offset) {
         if (StringUtils.isEmpty(tableName)) {
             throw new IllegalArgumentException("Table name cannot be null or empty");
         }
@@ -66,7 +66,7 @@ public class OracleDatabaseAdapter implements DatabaseAdapter {
         }
         if (nestedSelect) {
             query.append(") a");
-            int offsetVal = 0;
+            long offsetVal = 0;
             if (offset != null) {
                 offsetVal = offset;
             }
